+++ spec/simplecov_helper.rb
@@ -1,10 +1,22 @@
 # frozen_string_literal: true
 
 require "simplecov"
+require "simplecov-cobertura"
+require "simplecov-html"
+
+SimpleCov.formatters = [
+  SimpleCov::Formatter::CoberturaFormatter,
+  SimpleCov::Formatter::HTMLFormatter
+]
 
 SimpleCov.start do
   enable_coverage :branch
+
+  track_files "lib/cattri/**/*.rb"
+
+  add_filter "lib/cattri/version.rb"
   add_filter "/spec/"
+
   add_group "DataObjects", "lib/castkit/data_object"
   add_group "Attributes", "lib/castkit/attribute"
   add_group "Contracts", "lib/castkit/contract"
@@ -12,7 +24,4 @@ SimpleCov.start do
   add_group "Plugins", "lib/castkit/plugins"
 end
 
-SimpleCov.minimum_coverage 50
-
-require "simplecov-cobertura"
-SimpleCov.formatter = SimpleCov::Formatter::CoberturaFormatter
+SimpleCov.minimum_coverage 100